// apps/web/src/types/generated/CivilRefresh.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Snapshot } from "./Snapshot";
import type { StockSpec } from "./StockSpec";
import type { TickFrame } from "./TickFrame";

export type CivilRefresh = {
  ticks_per_day: number;
  snapshot: Snapshot;
  securities: Array<StockSpec>;
  intraday: Array<TickFrame>;
  public_publication_ids: Array<string>;
};