// apps/web/src/types/generated/ContinuousPoint.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Money } from "./Money";
import type { TradingPhase } from "./TradingPhase";

export type ContinuousPoint = {
  tick: number;
  phase: TradingPhase;
  last_price: Money;
  cumulative_volume: number;
  bids: Array<[Money, number]>;
  asks: Array<[Money, number]>;
};