// apps/web/src/types/generated/EventFact.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Event } from "./Event";
import type { EventStableKey } from "./EventStableKey";

export type EventFact = {
  key: EventStableKey;
  event: Event;
  canonical_payload: string;
};