// apps/web/src/types/generated/TickFrame.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Event } from "./Event";
import type { EventFact } from "./EventFact";
import type { TickTimeseriesPayload } from "./TickTimeseriesPayload";

export type TickFrame = {
  tick: number;
  events: Array<Event>;
  facts: Array<EventFact>;
  timeseries_payload: TickTimeseriesPayload;
  seq_from: number;
  seq_to: number;
};