// apps/web/src/types/generated/TickTimeseriesPayload.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { AuctionPoint } from "./AuctionPoint";
import type { ContinuousPoint } from "./ContinuousPoint";
import type { DailyCandle } from "./DailyCandle";
import type { MarketSnap } from "./MarketSnap";
import type { StockCode } from "./StockCode";

export type TickTimeseriesPayload = {
  markets: { [key in StockCode]: MarketSnap };
  active_daily_candles: { [key in StockCode]: DailyCandle };
  closed_daily_candles: { [key in StockCode]: DailyCandle };
  auction_points: { [key in StockCode]: Array<AuctionPoint> };
  continuous_points: { [key in StockCode]: ContinuousPoint };
};